feat: v1.0.0-beta.3.3 — auto-update via public manifest, obfuscation, file size fix
- Auto-update: rewrite GitHubUpdateChecker to read version.json from
  public bim-pills-releases repo (no auth required, private code repo unchanged)
- Auto-update: fix version parser for 1.0.0-beta.X.Y format (was broken for
  private repo tags — silently never detected updates)
- Auto-update: split HttpClient into API (10s) and download (30min) clients
- Auto-update: clean up ShowOverlayError dead code, add IsLoaded guard for
  race condition on async download
- Auto-update: create github.com/BIM-CA/bim-pills-releases public repo
- Build: add Obfuscar post-build step for Release — obfuscates Core,
  Infrastructure, Commands assemblies (KeepPublicApi=true)
- Build: bump version to 1.0.0-beta.3.3
- Audit: fix file size display for BIM 360/ACC collaborative models
  (CollaborationCache strategy — most recent .rvt by modification time)
- Audit: restore ISO references in methodology window (ISO 19650-1/2,
  AEC UK BIM Protocol, BIM Forum LOD)
- UI: UpdateAvailableWindow redesign with BIM Pills logo, download overlay,
  markdown cleaning in release notes, centered version badges
- UI: fix methodology window header style
- Sandbox: move Importar button to Panel Datos section

+++ src/BIMPills.Revit/Commands/Ordering/OrderingUndoHandler.cs
@@ -29,7 +29,7 @@ namespace BIMPills.Revit.Commands.Ordering
 
             try
             {
-                var element = doc.GetElement(new ElementId((int)entry.ElementId));
+                var element = doc.GetElement(new ElementId(entry.ElementId));
                 var param   = element?.LookupParameter(_session.Config.ParameterName);
 
                 if (param != null && !param.IsReadOnly)
